+++ src/main/java/ispd/gui/iconico/ruler/Ruler.java
@@ -28,7 +28,7 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package ispd.gui.iconico;
+package ispd.gui.iconico.ruler;
 
 import ispd.gui.utils.fonts.*;
 import java.awt.*;
